gh-658: refactor janus connector addClass
Luckily, most of the logic between addClass and updateClass was
duplicated, so we could reuse alomost all of what we already extensively
refactored for updateClass.
