+++ database/connectors/janusgraph/helper.go
@@ -16,213 +16,14 @@ import (
 	"errors"
 	"fmt"
 	"strings"
-	"time"
 
 	connutils "github.com/creativesoftwarefdn/weaviate/database/connectors/utils"
 	"github.com/creativesoftwarefdn/weaviate/database/schema"
 	"github.com/creativesoftwarefdn/weaviate/database/schema/kind"
 	"github.com/creativesoftwarefdn/weaviate/gremlin"
 	"github.com/creativesoftwarefdn/weaviate/models"
-	"github.com/davecgh/go-spew/spew"
 	"github.com/go-openapi/strfmt"
 )
-
-func (j *Janusgraph) addClass(k kind.Kind, className schema.ClassName, UUID strfmt.UUID, atContext string, creationTimeUnix int64, lastUpdateTimeUnix int64, key *models.SingleRef, rawProperties interface{}) error {
-	vertexLabel := j.state.getMappedClassName(className)
-
-	q := gremlin.G.AddV(string(vertexLabel)).
-		As("newClass").
-		StringProperty(PROP_KIND, k.Name()).
-		StringProperty(PROP_UUID, UUID.String()).
-		StringProperty(PROP_CLASS_ID, string(vertexLabel)).
-		StringProperty(PROP_AT_CONTEXT, atContext).
-		Int64Property(PROP_CREATION_TIME_UNIX, creationTimeUnix).
-		Int64Property(PROP_LAST_UPDATE_TIME_UNIX, lastUpdateTimeUnix)
-
-	// map properties in thing.Schema according to the mapping.
-	type edgeToAdd struct {
-		PropertyName string
-		Type         string
-		Reference    string
-		Location     string
-	}
-
-	var edgesToAdd []edgeToAdd
-
-	properties, schema_ok := rawProperties.(map[string]interface{})
-
-	if schema_ok {
-		for propName, value := range properties {
-			sanitizedPropertyName := schema.AssertValidPropertyName(propName)
-			err, property := j.schema.GetProperty(k, className, sanitizedPropertyName)
-			if err != nil {
-				return err
-			}
-
-			janusPropertyName := string(j.state.getMappedPropertyName(className, sanitizedPropertyName))
-
-			propType, err := j.schema.FindPropertyDataType(property.AtDataType)
-			if err != nil {
-				return err
-			}
-
-			if propType.IsPrimitive() {
-				switch propType.AsPrimitive() {
-				case schema.DataTypeInt:
-					switch t := value.(type) {
-					case int:
-						q = q.Int64Property(janusPropertyName, int64(t))
-					case int8:
-						q = q.Int64Property(janusPropertyName, int64(t))
-					case int16:
-						q = q.Int64Property(janusPropertyName, int64(t))
-					case int32:
-						q = q.Int64Property(janusPropertyName, int64(t))
-					case int64:
-						q = q.Int64Property(janusPropertyName, t)
-					case uint:
-						q = q.Int64Property(janusPropertyName, int64(t))
-					case uint8:
-						q = q.Int64Property(janusPropertyName, int64(t))
-					case uint16:
-						q = q.Int64Property(janusPropertyName, int64(t))
-					case uint32:
-						q = q.Int64Property(janusPropertyName, int64(t))
-					case uint64:
-						q = q.Int64Property(janusPropertyName, int64(t))
-					case float32:
-						q = q.Int64Property(janusPropertyName, int64(t))
-					case float64:
-						q = q.Int64Property(janusPropertyName, int64(t))
-					default:
-						return fmt.Errorf("Illegal value for property %s", sanitizedPropertyName)
-					}
-				case schema.DataTypeString:
-					switch t := value.(type) {
-					case string:
-						q = q.StringProperty(janusPropertyName, t)
-					default:
-						return fmt.Errorf("Illegal value for property %s", sanitizedPropertyName)
-					}
-				case schema.DataTypeText:
-					switch t := value.(type) {
-					case string:
-						q = q.StringProperty(janusPropertyName, t)
-					default:
-						return fmt.Errorf("Illegal value for property %s", sanitizedPropertyName)
-					}
-				case schema.DataTypeBoolean:
-					switch t := value.(type) {
-					case bool:
-						q = q.BoolProperty(janusPropertyName, t)
-					default:
-						return fmt.Errorf("Illegal value for property %s", sanitizedPropertyName)
-					}
-				case schema.DataTypeNumber:
-					switch t := value.(type) {
-					case float32:
-						q = q.Float64Property(janusPropertyName, float64(t))
-					case float64:
-						q = q.Float64Property(janusPropertyName, t)
-					default:
-						return fmt.Errorf("Illegal value for property %s", sanitizedPropertyName)
-					}
-				case schema.DataTypeDate:
-					switch t := value.(type) {
-					case time.Time:
-						q = q.StringProperty(janusPropertyName, t.Format(time.RFC3339))
-					default:
-						return fmt.Errorf("Illegal value for property %s", sanitizedPropertyName)
-					}
-				default:
-					panic(fmt.Sprintf("Unkown primitive datatype %s", propType.AsPrimitive()))
-				}
-			} else {
-				switch schema.CardinalityOfProperty(property) {
-				case schema.CardinalityAtMostOne:
-					switch t := value.(type) {
-					case *models.SingleRef:
-						var refClassName schema.ClassName
-						switch t.Type {
-						case "Action":
-							var singleRefValue models.ActionGetResponse
-							err = j.GetAction(nil, t.NrDollarCref, &singleRefValue)
-							if err != nil {
-								return fmt.Errorf("Illegal value for property %s; could not resolve action with UUID: %v", t.NrDollarCref.String(), err)
-							}
-							refClassName = schema.AssertValidClassName(singleRefValue.AtClass)
-						case "Thing":
-							var singleRefValue models.ThingGetResponse
-							err = j.GetThing(nil, t.NrDollarCref, &singleRefValue)
-							if err != nil {
-								return fmt.Errorf("Illegal value for property %s; could not resolve thing with UUID: %v", t.NrDollarCref.String(), err)
-							}
-							refClassName = schema.AssertValidClassName(singleRefValue.AtClass)
-						default:
-							return fmt.Errorf("Illegal value for property %s; only Thing or Action supported, got ", t.Type)
-						}
-
-						// Verify the cross reference
-						if !propType.ContainsClass(refClassName) {
-							return fmt.Errorf("Illegal value for property %s; cannot point to %s", sanitizedPropertyName, t.Type)
-						}
-						edgesToAdd = append(edgesToAdd, edgeToAdd{
-							PropertyName: janusPropertyName,
-							Reference:    t.NrDollarCref.String(),
-							Type:         t.Type,
-							Location:     *t.LocationURL,
-						})
-					default:
-						return fmt.Errorf("Illegal value for property %s", sanitizedPropertyName)
-					}
-				case schema.CardinalityMany:
-					switch t := value.(type) {
-					case models.MultipleRef:
-						for _, ref := range t {
-							edgesToAdd = append(edgesToAdd, edgeToAdd{
-								PropertyName: janusPropertyName,
-								Reference:    ref.NrDollarCref.String(),
-								Type:         ref.Type,
-								Location:     *ref.LocationURL,
-							})
-						}
-					default:
-						return fmt.Errorf("Illegal value for property %s", sanitizedPropertyName)
-					}
-				default:
-					panic(fmt.Sprintf("Unexpected cardinality %v", schema.CardinalityOfProperty(property)))
-				}
-			}
-		}
-	}
-	// Add edges to all referened things.
-	for _, edge := range edgesToAdd {
-		q = q.AddE(edge.PropertyName).
-			FromRef("newClass").
-			ToQuery(gremlin.G.V().HasString(PROP_UUID, edge.Reference)).
-			StringProperty(PROP_REF_ID, edge.PropertyName).
-			StringProperty(PROP_REF_EDGE_CREF, edge.Reference).
-			StringProperty(PROP_REF_EDGE_TYPE, edge.Type).
-			StringProperty(PROP_REF_EDGE_LOCATION, edge.Location)
-	}
-
-	// Link to key
-	q = q.AddE(KEY_VERTEX_LABEL).
-		StringProperty(PROP_REF_EDGE_CREF, string(key.NrDollarCref)).
-		StringProperty(PROP_REF_EDGE_TYPE, key.Type).
-		StringProperty(PROP_REF_EDGE_LOCATION, *key.LocationURL)
-
-	q = q.FromRef("newClass").
-		ToQuery(gremlin.G.V().HasLabel(KEY_VERTEX_LABEL).HasString(PROP_UUID, key.NrDollarCref.String()))
-
-	fmt.Print("\n\n\n\n\n\n\n\n")
-	spew.Dump(q)
-	fmt.Print("\n\n\n\n\n\n\n\n")
-
-	_, err := j.client.Execute(q)
-
-	return err
-}
 
 func (j *Janusgraph) getClass(k kind.Kind, searchUUID strfmt.UUID, atClass *string, atContext *string, foundUUID *strfmt.UUID, creationTimeUnix *int64, lastUpdateTimeUnix *int64, properties *models.Schema, key **models.SingleRef) error {
 	// Fetch the class, it's key, and it's relations.
