+++ database/connectors/janusgraph/helper_add_class.go
@@ -1,0 +1,41 @@
+package janusgraph
+
+import (
+	"github.com/creativesoftwarefdn/weaviate/database/schema"
+	"github.com/creativesoftwarefdn/weaviate/database/schema/kind"
+	"github.com/creativesoftwarefdn/weaviate/gremlin"
+	"github.com/creativesoftwarefdn/weaviate/models"
+	"github.com/go-openapi/strfmt"
+)
+
+func (j *Janusgraph) addClass(k kind.Kind, className schema.ClassName, UUID strfmt.UUID, atContext string, creationTimeUnix int64, lastUpdateTimeUnix int64, key *models.SingleRef, rawProperties interface{}) error {
+	vertexLabel := j.state.getMappedClassName(className)
+	sourceClassAlias := "classToBeAdded"
+
+	q := gremlin.G.AddV(string(vertexLabel)).
+		As(sourceClassAlias).
+		StringProperty(PROP_KIND, k.Name()).
+		StringProperty(PROP_UUID, UUID.String()).
+		StringProperty(PROP_CLASS_ID, string(vertexLabel)).
+		StringProperty(PROP_AT_CONTEXT, atContext).
+		Int64Property(PROP_CREATION_TIME_UNIX, creationTimeUnix).
+		Int64Property(PROP_LAST_UPDATE_TIME_UNIX, lastUpdateTimeUnix)
+
+	q, err := j.addEdgesToQuery(q, k, className, rawProperties, sourceClassAlias)
+	if err != nil {
+		return err
+	}
+
+	// Link to key
+	q = q.AddE(KEY_VERTEX_LABEL).
+		StringProperty(PROP_REF_EDGE_CREF, string(key.NrDollarCref)).
+		StringProperty(PROP_REF_EDGE_TYPE, key.Type).
+		StringProperty(PROP_REF_EDGE_LOCATION, *key.LocationURL)
+
+	q = q.FromRef(sourceClassAlias).
+		ToQuery(gremlin.G.V().HasLabel(KEY_VERTEX_LABEL).HasString(PROP_UUID, key.NrDollarCref.String()))
+
+	_, err = j.client.Execute(q)
+
+	return err
+}
